+++ geo-types/src/line.rs
@@ -1,3 +1,5 @@
+#[cfg(any(feature = "rstar_0_8", feature = "rstar_0_9"))]
+use crate::CoordFloat;
 use crate::{CoordNum, Coordinate, Point};
 #[cfg(any(feature = "approx", test))]
 use approx::{AbsDiffEq, RelativeEq};
@@ -226,7 +228,7 @@ macro_rules! impl_rstar_line {
     ($rstar:ident) => {
         impl<T> ::$rstar::RTreeObject for Line<T>
         where
-            T: ::num_traits::Float + ::$rstar::RTreeNum,
+            T: CoordFloat + ::$rstar::RTreeNum,
         {
             type Envelope = ::$rstar::AABB<Point<T>>;
 
@@ -238,7 +240,7 @@ macro_rules! impl_rstar_line {
 
         impl<T> ::$rstar::PointDistance for Line<T>
         where
-            T: ::num_traits::Float + ::$rstar::RTreeNum,
+            T: CoordFloat + ::$rstar::RTreeNum,
         {
             fn distance_2(&self, point: &Point<T>) -> T {
                 let d = crate::private_utils::point_line_euclidean_distance(*point, *self);