+++ geo/CHANGES.md
@@ -5,6 +5,7 @@
 * Add `LinesIter` algorithm to iterate over the lines in geometries.
   * Very similar to `CoordsIter`, but only implemented where it makes sense (e.g., for `Polygon`, `Rect`, but not `Point`).
   * <https://github.com/georust/geo/pull/757>
+* BREAKING: Add `Default` constraint to all `CoordNum`/`CoordFloat` values.
 
 ## 0.19.0